+++ src/helper.ts
@@ -3,7 +3,9 @@ import * as vscode from 'vscode';
 export interface IFormat {
     format: string,
     extensionName: string,
-    savePath: string
+    savePath: string,
+    savePathSegmentKeys: string[] | null,
+    savePathReplaceSegmentsWith: string
 }
 
 export class Helper {